+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.3.10-alpha.8 (2021-04-05)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
+### 2.3.10-alpha.7 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
+### 2.3.10-alpha.6 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-dom
+
+
+
+
+
 ### 2.3.10-alpha.1 (2021-04-01)
 
 **Note:** Version bump only for package react-magma-dom