+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.0.3-alpha.9 (2021-04-05)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 2.0.3-alpha.8 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 2.0.3-alpha.7 (2021-04-02)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
 ### 2.0.3-alpha.6 (2021-04-01)
 
 **Note:** Version bump only for package react-magma-icons